Great readme just fixing a few typos
The readme looks very nice I am just correcting a few typos. For the segments where I needed to think to decipher what was written or there was a lot to do, I used ChatGPT to correct it and then checked if the correction was alright.

Note: I still want to implement some of Mathematica's graph/network functions but I  am caught up in a bunch of other stuff.